Problem: Kwery used to use an outdated library https://github.com/xebialabs/overcast for testing databases through docker.
Solution: Moving to a better and newer library https://www.testcontainers.org/. This has containers for MySQL, PostgreSQL out of the box which internally handles container boot up wait etc.

Note: Has a lot of import optimisation due to which almost the entire source code is part of this check-in.