Problem: Test cases which used SQL queries were failing.
Solution: This is happening because earlier the database bootstrapping through containers was done through a different library. To simplify this using SQL queries for testing which do not depend on any tables being present.

Note: A lot of UI test cases are failing, this is due to brittle test cases and not due to bugs in the product. Need to tackle this later.